+++ test/main.cpp
@@ -24,7 +24,7 @@
 
 int main(int argc, char *argv[])
 {
-    hdbpp_internal::LogConfigurator::initLogging(false, true, "/tmp/hdb/test.log");
+    hdbpp_internal::LogConfigurator::initLogging();
     hdbpp_internal::LogConfigurator::setLoggingLevel(spdlog::level::err);
 
     int result = Catch::Session().run(argc, argv);